Condensing the GLForm
